+++ src/test/java/seedu/fridget/FridgetTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.fridget;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class FridgetTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
